% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecars.R
\name{ecars}
\alias{ecars}
\title{Calculate importance}
\usage{
ecars(X)
}
\arguments{
\item{X}{A nxp data matrix}
}
\value{
A vector of each variable's importance value
}
\description{
Assign importance value to each variable in the
regression model.
}
\examples{
ecars(X)
}
